multidatabase/MultiDB_Django/requirements.txt
asgiref==3.5.2
cffi==1.15.1
configparser==5.3.0
cryptography==38.0.3
Django==4.1
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.13.1
mysql-connector-python==8.0.31
nmap==0.0.1
numpy==1.23.5
packaging==21.3
pandas==1.5.1
protobuf==3.20.1
psycopg2==2.9.5
pyasn1==0.4.8
pycparser==2.21
pyexasol==0.25.1
pyodbc==4.0.34
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2
rsa==4.9
six==1.16.0
sqlparse==0.4.2
websocket-client==1.4.2
